Fixed compiler warnings

+++ include/passive-gripper/models/MeshDependentResource.cpp
@@ -73,9 +73,9 @@ void MeshDependentResource::init_sp() const {
     bool operator<(const VertexInfo& r) const { return dist > r.dist; }
   };
 #pragma omp parallel for
-  for (int src = 0; src < nV; src++) {
+  for (size_t src = 0; src < nV; src++) {
     std::priority_queue<VertexInfo> q;
-    q.push(VertexInfo{src, 0});
+    q.push(VertexInfo{(int) src, 0});
     SP_(src, src) = 0;
     while (!q.empty()) {
       VertexInfo u = q.top();
